refactor#61: sharedPreference 및 enum 관련 코드 리팩토링
- sharedPreference의 경우 DataStore로 교체 필요 #65
